requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-filter==23.5
psycopg2-binary==2.9.9
python-multipart==0.0.6
openpyxl==3.1.2
pandas==2.1.4
PyPDF2==3.0.1
python-magic==0.4.27
pillow==10.1.0
dj-database-url==2.1.0
python-dotenv==1.0.0
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
drf-yasg==1.21.7
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
threading-timer==0.1.0
dj-database-url
djangorestframework